fix: Convert to float32 for MSE loss calculation to support BFloat16
- MSE loss not implemented for BFloat16 on CPU
- Convert tensors to float32 only for loss computation
- Preserves BFloat16 throughout model forward pass
- Fixes RuntimeError: "mse_cpu" not implemented for 'BFloat16'
- Stage 1 adapter training should now run successfully on H100s

+++ train_adapter_only.py
@@ -224,9 +224,10 @@ def train_adapter_only(args):
                 reconstructed = reconstructed.to(orig_embeds.dtype)
 
             # Loss 1: Reconstruction
+            # Convert to float32 for MSE loss (not supported for BFloat16)
             recon_loss = F.mse_loss(
-                reconstructed[attention_mask.bool()],
-                orig_embeds[attention_mask.bool()]
+                reconstructed[attention_mask.bool()].float(),
+                orig_embeds[attention_mask.bool()].float()
             )
 
             # Loss 2: Generation quality
